whitespace consistency across docs
I saw a rendedered whitespace inconsistency in the Tasks docs here:
https://github.com/joaomdmoura/crewAI/blob/ed318600719f07721240f76f5024c6d45f837045/docs/core-concepts/Tasks.md?plain=1#L173

So I set out to patch that up to make it easier to read.  I then noticed
there were a few whitespace inconsistencies:
- 2 spaces
- 4 whitespaces
- tabs

It appears that the 4 whitespaces is the prevalent whitesapce usage, so
I overwrote other whitespace usages with that in this commit.